Servercode/requirements.txt
fastapi==0.115.2
Pillow==10.4.0
torch==2.4.1
torchvision==0.19.1
uvicorn==0.31.1
